customer_chatbot/requirements.txt
groq>=0.4.0
requests>=2.31.0
streamlit>=1.29.0
python-dotenv>=1.0.0
